% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_graph.R
\name{g3_graph_amp}
\alias{g3_graph_amp}
\title{Plot amplification graph}
\usage{
g3_graph_amp(g3obj)
}
\arguments{
\item{g3obj}{g3 object}
}
\value{
returns graph of amplification curves
}
\description{
Plot amplification graph
}
