% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_graph.R
\name{g3_graph_amp_point}
\alias{g3_graph_amp_point}
\title{Plot amplification scatter}
\usage{
g3_graph_amp_point(g3obj)
}
\arguments{
\item{g3obj}{g3 object}
}
\value{
returns scatter of amplification in a ploty-compatible form
}
\description{
Plot amplification scatter
}
