% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_graph.R
\name{g3_graph_anneal}
\alias{g3_graph_anneal}
\title{Plot annealing graph}
\usage{
g3_graph_anneal(g3obj)
}
\arguments{
\item{g3obj}{g3 object}
}
\value{
returns ggplot graph of annealing curves
}
\description{
Plot annealing graph
}
