% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_read.R
\name{g3_read}
\alias{g3_read}
\title{Read Genie-III csv trio}
\usage{
g3_read(g3index)
}
\arguments{
\item{g3index}{character. Filepath and 'common' filename of the Genie-III csv trio. Should NOT include
either the .CSV ending or any fileendings after and including a - symbol (so if file is X/Gen123-Anneal),
would input 'X/Gen123'}
}
\value{
returns a list containing key information about the Genie-III experiment run and tibbles covering
summary, anneal, and amplification data.
}
\description{
Takes character file location & name of a Genie csv set (minus the endings) and returns a list of
summary, anneal, and amplification data for the object, ready for further processing
}
