% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_read_amp.R
\name{g3_read_amp}
\alias{g3_read_amp}
\title{Read Genie-III csv summary + amplification file duo}
\usage{
g3_read_amp(g3index)
}
\arguments{
\item{g3index}{character. Filepath and 'common' filename of the Genie-III csv trio. Should NOT include
either the .CSV ending or any fileendings after and including a - symbol (so if file is X/Gen123-Amp),
would input 'X/Gen123'}
}
\value{
returns a list containing key information about the Genie-III experiment run and tibbles covering
summary, and amplification data.
}
\description{
Takes character file location & name of a Genie csv set (minus the endings) and returns a list of
summary, and amplification data for the object, ready for further processing
}
\examples{
g3_read_amp("test")
}
