% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_read_anneal.R
\name{g3_read_anneal}
\alias{g3_read_anneal}
\title{Read Genie-III csv summary + anneal file duo}
\usage{
g3_read_anneal(g3index)
}
\arguments{
\item{g3index}{character. Filepath and 'common' filename of the Genie-III csv trio. Should NOT include
either the .CSV ending or any fileendings after and including a - symbol (so if file is X/Gen123-Anneal),
would input 'X/Gen123'}
}
\value{
returns a list containing key information about the Genie-III experiment run and tibbles covering
summary and anneal data.
}
\description{
Takes character file location & name of a Genie csv set (minus the endings) and returns a list of
summary and anneal data for the object, ready for further processing
}
