% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_read_summary.R
\name{g3_read_summary}
\alias{g3_read_summary}
\title{Read Genie-III csv summary file}
\usage{
g3_read_summary(g3index)
}
\arguments{
\item{g3index}{character. Filepath and 'common' filename of the Genie-III csv trio. Should NOT include
either the .CSV ending}
}
\value{
returns a list containing key information about the Genie-III experiment run and tibbles covering
the summary data.
}
\description{
Takes character file location & name of a Genie csv set (minus the endings) and returns
summary data for the object, ready for further processing
}
