% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_summarise.R
\name{g3_summarise}
\alias{g3_summarise}
\title{Summarise Tt information for g3 object}
\usage{
g3_summarise(g3obj, threshold = 20000)
}
\arguments{
\item{g3obj}{g3 object}

\item{threshold}{fluorescence threshold that must be exceeded}
}
\value{
returns tibble showing time eligible rows exceeded the threshold
}
\description{
Summarise Tt information for g3 object
}
\examples{
g3_load("test") \%>\% g3_summarise()
}
